improve test split and coverage
