+++ scripts/test/Engineering/texture/TextureUtils/CMakeLists.txt
@@ -1,7 +1,13 @@
 # Texture Util tests
 
-set(TEST_PY_FILES test_io.py test_correction_utils.py test_focus_utils.py test_fitting_utils.py
-                  test_polefigure_utils.py
+set(TEST_PY_FILES
+    test_io.py
+    test_correction_utils.py
+    test_focus_utils.py
+    test_fitting_utils.py
+    test_fitpeaks_engine.py
+    test_multidomain_engine.py
+    test_polefigure_utils.py
 )
 
 check_tests_valid(${CMAKE_CURRENT_SOURCE_DIR} ${TEST_PY_FILES})
